[refactor] 코드 리팩토링
- empty 찾는 조건 리팩토링
- 게시글 서비스 중 사용자 서비스와 관련된 부분은 UserService 로 이동
