+++ src/main/java/org/board/exercise_board/Post/service/PostService.java
@@ -1,7 +1,7 @@
 package org.board.exercise_board.Post.service;
 
-import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.board.exercise_board.Post.domain.Dto.PostDto;
 import org.board.exercise_board.Post.domain.form.ModifyForm;
@@ -11,7 +11,7 @@ import org.board.exercise_board.Post.domain.repository.PostRepository;
 import org.board.exercise_board.Post.exception.PostCustomException;
 import org.board.exercise_board.Post.exception.PostErrorCode;
 import org.board.exercise_board.User.domain.model.User;
-import org.board.exercise_board.User.domain.repository.UserRepository;
+import org.board.exercise_board.User.service.UserService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -22,7 +22,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class PostService {
 
   private final PostRepository postRepository;
-  private final UserRepository userRepository;
+  private final UserService userService;
 
   /**
    * 해당 사용자 정보와 입력 받은 Post 제목, 내용을 함께 DB에 저장
@@ -32,7 +32,7 @@ public class PostService {
    * @return
    */
   public PostDto savePost(WriteForm writeForm, String writerId) {
-    User user = findUser(writerId);
+    User user = userService.findUser(writerId);
 
     Post post = Post.formToEntity(writeForm);
     post.setUser(user);
@@ -40,16 +40,7 @@ public class PostService {
     return PostDto.entityToDto(postRepository.save(post));
   }
 
-  /**
-   * 해당 사용자가 이메일 인증을 받았는지 검사
-   *
-   * @param writerId
-   * @return
-   */
-  public boolean isEmailVerified(String writerId) {
-    User user = findUser(writerId);
-    return user.getVerifiedStatus();
-  }
+
 
   /**
    * 사용자의 ID와 제목으로 해당하는 게시물 중 첫번째 게시물 찾기
@@ -59,7 +50,7 @@ public class PostService {
    * @return
    */
   public Post findPosts(String writerId, String subject) {
-    User user = findUser(writerId);
+    User user = userService.findUser(writerId);
 
     // 해당 게시물이 존재하지 않는다면 에러 발생
     if (postRepository.findAllBySubjectAndUser(subject, user).isEmpty()) {
@@ -76,18 +67,6 @@ public class PostService {
   }
 
 
-  /**
-   * ID로 해당 사용자 찾는 메서드
-   *
-   * @param writerId
-   * @return
-   */
-  public User findUser(String writerId) {
-    User user = userRepository.findByLoginId(writerId)
-        .orElseThrow(() -> new PostCustomException(PostErrorCode.NOT_FOUND_USER));
-
-    return user;
-  }
 
   /**
    * 최근 날짜 순으로 모든 게시물 조회(Entity -> DTO)
@@ -113,13 +92,9 @@ public class PostService {
    */
   @Transactional
   public List<PostDto> searchPost(String keyword, Pageable pageable) {
-    Page<Post> postLists = postRepository.findBySubjectContaining(keyword, pageable);
 
-    List<PostDto> postDtos = new ArrayList<>();
-    for (Post post : postLists) {
-      postDtos.add(PostDto.entityToDto(post));
-    }
-
-    return postDtos;
+    return postRepository.findBySubjectContaining(keyword,pageable)
+        .stream().map(it-> PostDto.entityToDto(it))
+        .collect(Collectors.toList());
   }
 }
