+++ src/main/java/org/board/exercise_board/User/service/UserService.java
@@ -1,8 +1,9 @@
 package org.board.exercise_board.User.service;
 
 import lombok.RequiredArgsConstructor;
+import org.board.exercise_board.Post.exception.PostCustomException;
+import org.board.exercise_board.Post.exception.PostErrorCode;
 import org.board.exercise_board.User.Security.JwtTokenProvider;
-import org.board.exercise_board.User.domain.Dto.UserDto;
 import org.board.exercise_board.User.domain.Form.SignUpForm;
 import org.board.exercise_board.User.domain.model.JwtToken;
 import org.board.exercise_board.User.domain.model.User;
@@ -71,5 +72,30 @@ public class UserService {
     JwtToken token = jwtTokenProvider.createToken(authentication);
 
     return token;
+  }
+
+  /**
+   * 해당 사용자가 이메일 인증을 받았는지 검사
+   *
+   * @param writerId
+   * @return
+   */
+  public boolean isEmailVerified(String writerId) {
+    User user = userRepository.findByLoginId(writerId)
+        .orElseThrow(() -> new PostCustomException(PostErrorCode.NOT_FOUND_USER));
+    return user.getVerifiedStatus();
+  }
+
+  /**
+   * ID로 해당 사용자 찾는 메서드
+   *
+   * @param writerId
+   * @return
+   */
+  public User findUser(String writerId) {
+    User user = userRepository.findByLoginId(writerId)
+        .orElseThrow(() -> new PostCustomException(PostErrorCode.NOT_FOUND_USER));
+
+    return user;
   }
 }
